requirements.txt
Django==2.1.7
django-crispy-forms==1.7.2
django-environ==0.4.5
django-markdownx==2.0.28
Markdown==3.0.1
Pillow==5.4.1
pytz==2018.9
